@@ -1,10 +1,11 @@
 # frozen_string_literal: true
+
 # This file is copied to spec/ when you run 'rails generate rspec:install'
 ENV['RAILS_ENV'] ||= 'test'
 require 'spec_helper'
-require File.expand_path('../../config/environment', __FILE__)
+require File.expand_path('../config/environment', __dir__)
 require 'rspec/rails'
-require 'factory_girl_rails'
+require 'factory_bot_rails'
 Dir['./spec/support/**/*.rb'].sort.each { |f| require f }
 # Add additional requires below this line. Rails is not loaded until this point!
 
@@ -28,7 +29,7 @@ Dir['./spec/support/**/*.rb'].sort.each { |f| require f }
 ActiveRecord::Migration.maintain_test_schema!
 
 RSpec.configure do |config|
-  config.include FactoryGirl::Syntax::Methods
+  config.include FactoryBot::Syntax::Methods
   config.include Devise::Test::ControllerHelpers, type: :controller
   config.extend  ControllerMacros, type: :controller
 
