@@ -1,8 +1,9 @@
 # frozen_string_literal: true
+
 class LeaveTimesController < BaseController
   STARTING_YEAR = Settings.misc.starting_year.to_i
   SHOWINGS = %i(all effective).freeze
-  DEFAULT_SHOWING = 'effective'.freeze
+  DEFAULT_SHOWING = 'effective'
   before_action :set_query_object
 
   helper_method :showing
