@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 module LeaveApplicationsHelper
   def time_object_by_hour_min(hour, min)
     Time.zone.now.change(hour: hour, min: min)
@@ -25,25 +26,25 @@ module LeaveApplicationsHelper
   end
 
   def need_deduct_salary?(leave_type, leave_time)
-    return if leave_time == 0
-    leave_type == 'personal' || leave_type == 'halfpaid_sick'
+    return if leave_time.zero?
+
+    %w(personal halfpaid_sick).include? leave_type
   end
 
   def summary_from(summary, user_id)
     leave_types = Settings.leave_times.quota_types.keys
-    summary[user_id] || Hash[leave_types.collect{ |type| [type, 0]}]
+    summary[user_id] || Hash[leave_types.collect { |type| [type, 0] }]
   end
 
   def leave_type_dropdown_menu(action_name, user)
     if user.role == 'contractor'
-      LeaveApplication.enum_attributes_for_select(:leave_types, except = [:remote, :sick, :maternity, :marriage, :compassionate, :official, :occpational_sick, :menstrual])
+      LeaveApplication.enum_attributes_for_select(:leave_types, [:remote, :sick, :maternity, :marriage, :compassionate, :official, :occpational_sick, :menstrual])
     else
       LeaveApplication.enum_attributes_for_select(:leave_types)
     end
   end
 
   def link_to_attachment(attachment)
-    link_to t('.exam_attachment'), current_object.attachment.url, target: :_blank if attachment
+    link_to t('.exam_attachment'), current_object.attachment.url, target: :_blank, rel: 'noopener' if attachment
   end
-
 end
