@@ -1,8 +1,9 @@
 # frozen_string_literal: true
+
 module MetaDataHelper
   def page_title
     "#{t(page_title_translation_key, raise: true)} | #{t('misc.app_title')}"
-  rescue
+  rescue I18n::MissingTranslationData
     t('misc.app_title')
   end
 
