@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 class LeaveApplication < ApplicationRecord
   include AASM
   include SignatureConcern
@@ -14,9 +15,9 @@ class LeaveApplication < ApplicationRecord
   paginates_per 15
 
   belongs_to :user
-  belongs_to :manager, class_name: 'User', foreign_key: 'manager_id'
+  belongs_to :manager, class_name: 'User', foreign_key: 'manager_id', inverse_of: :leave_applications
+  has_many   :leave_time_usages, dependent: :delete_all
   has_many   :leave_times, through: :leave_time_usages
-  has_many   :leave_time_usages
   has_many   :leave_hours_by_dates, dependent: :delete_all
 
   validates :leave_type, :description, :start_time, :end_time, presence: true
@@ -88,6 +89,7 @@ class LeaveApplication < ApplicationRecord
 
   def check_special_leave_application_quota
     return true unless self.special_type?
+
     self.leave_time_usages.any? or self.available_leave_times.pluck(:usable_hours).sum >= self.hours
   end
 
@@ -135,19 +137,23 @@ class LeaveApplication < ApplicationRecord
 
   def auto_calculated_minutes
     return @minutes = 0 unless start_time && end_time
+
     @minutes = Daikichi::Config::Biz.within(start_time, end_time).in_minutes
   end
 
   def hours_should_be_positive_integer
     return if self.errors[:start_time].any? or self.errors[:end_time].any?
+
     errors.add(:end_time, :not_integer) if (@minutes % 60).nonzero? || !self.hours.positive?
     errors.add(:start_time, :should_be_earlier) unless self.end_time > self.start_time
   end
 
   def should_not_overlaps_other_applications
     return if self.errors[:start_time].any? or self.errors[:end_time].any?
+
     overlapped = LeaveApplication.personal(user_id, start_time, end_time).where.not(id: self.id)
     return unless overlapped.any?
+
     overlap_application_error_messages(overlapped)
   end
 
@@ -160,14 +166,15 @@ class LeaveApplication < ApplicationRecord
           'activerecord.errors.models.leave_application.attributes.base.overlap_application',
           leave_type: LeaveApplication.human_enum_value(:leave_type, la.leave_type),
           start_time: la.start_time.to_formatted_s(:month_date),
-          end_time:   la.end_time.to_formatted_s(:month_date),
-          link:       url.leave_application_path(id: la.id)
+          end_time: la.end_time.to_formatted_s(:month_date),
+          link: url.leave_application_path(id: la.id)
         )
       )
     end
   end
 
   def order_by_sequence
-    format 'array_position(Array%s, leave_type::TEXT)', Settings.leave_applications.available_quota_types.send(self.leave_type).to_s.tr('"', "'")
+    leavetype = Settings.leave_applications.available_quota_types.send(self.leave_type)
+    Arel.sql "array_position(Array['#{leavetype}'], leave_type::TEXT)"
   end
 end
