@@ -1,23 +1,24 @@
 # frozen_string_literal: true
+
 class LeaveTime < ApplicationRecord
   delegate :seniority, :name, to: :user
 
   enum leave_type: Settings.leave_times.quota_types
 
   belongs_to :user
-  has_many   :leave_applications, through: :leave_time_usages
-  has_many   :leave_time_usages
+  has_many :leave_time_usages, dependent: :delete_all
+  has_many :leave_applications, through: :leave_time_usages
 
   before_validation :set_default_values
   after_create :build_special_leave_time_usages
 
   validates :leave_type, :effective_date, :expiration_date, :quota, :usable_hours, :used_hours, :locked_hours, :user, presence: true
-  validates :quota,        numericality: { only_integer: true, greater_than_or_equal_to: 0 }
+  validates :quota, numericality: { only_integer: true, greater_than_or_equal_to: 0 }
   validates :usable_hours, numericality: { only_integer: true, greater_than_or_equal_to: 0 }
-  validates :used_hours,   numericality: { only_integer: true, greater_than_or_equal_to: 0 }
+  validates :used_hours, numericality: { only_integer: true, greater_than_or_equal_to: 0 }
   validates :locked_hours, numericality: { only_integer: true, greater_than_or_equal_to: 0 }
-  validate  :positive_range
-  validate  :balanced_hours
+  validate :positive_range
+  validate :balanced_hours
 
   scope :belong_to, ->(user) {
     where(user: user)
@@ -128,6 +129,7 @@ class LeaveTime < ApplicationRecord
   def positive_range
     return if self.errors[:effective_date].any? || self.errors[:expiration_date].any?
     return if expiration_date >= effective_date
+
     errors.add(:effective_date, :range_should_be_positive)
   end
 
@@ -139,11 +141,13 @@ class LeaveTime < ApplicationRecord
 
   def balanced_hours
     return if errors[:usable_hours].any? or errors[:used_hours].any? or errors[:locked_hours].any?
+
     errors.add(:quota, :unbalanced_hours) if quota != (usable_hours + used_hours + locked_hours)
   end
 
   def build_special_leave_time_usages
     return unless self.special_type?
+
     leave_applications = User.find(self.user_id).leave_applications.where(leave_type: self.leave_type)
     leave_applications.each do |la|
       if la.leave_time_usages.empty? and la.hours <= la.available_leave_times.pluck(:usable_hours).sum
