@@ -1,5 +1,6 @@
 # frozen_string_literal: true
+
 class Settings < Settingslogic
-  source "#{Rails.root}/config/application.yml"
+  source Rails.root.join('config', 'application.yml')
   namespace Rails.env
 end
